Docu-bonanza, spring cleaning of base types.
